requirements.txt
flask-restplus==0.12.1
Flask-Cors==3.0.8
Flask-Login==0.4.1
PyJWT==1.7.1
uWSGI==2.0.17.1
Flask==1.0.4
Flask-Migrate==2.5.2
flask-restplus==0.12.1
Flask-SQLAlchemy==2.4.1
Flask-HTTPAuth==3.3.0
psycopg2==2.8.3
py-money==0.4.0
numpy==1.18.0
numpy-financial==1.0.0
json-logging==1.0.5
newrelic==5.6.0.135
